Fix SVG on Safari
For some reason Safari was not liking loading SVG from data urls. This
imports them and pulls the SVG inline which avoids this problem without
having to repeat or genuinely hard code SVGs

Added a new plugin and set it to only allow us of SVGs with the
.inline.svg extension.

Most stuff seems to be working.

PNG badges on the talks page still seem problematic. Will solve that
separately.