Work in progress giving form
This is intended to allow people to respond to upcoming financial update
without having to use paper.

I'm sure there's plenty of refactoring that can be done here. Especially
if we want to use similar functionality for other forms.

This basically forms a URL to submit data to a Google Form and hides the
nasty google forms UI.

Submitting this for some feedback