Allow development on giving form
When development environment is detected (i.e. local dev on `yarn
start`) this will use development form config and show a nice red box.

This will send form data to Tom's form. Ask me for access if you need
it.

Added typing for the config too.

Moved the google form id out of the markdown to make this a bit easier.