Allow color scheme override for sections
Not all sections are dark or light. London Living is a prime example.

Provide the means to override this option by moving to an enum based system.

"custom" setting avoids setting and background colors.

There must be a better way overall to do this but this is at least a bit better.