Adding covid 19 policy and risk assessment docs
This document is something that we must have avaiable for public viewing.

I have used html tables as MD tables were not able to do spanrows/spancolumns and alternatives where way too hard to work with.

In this case I used an online converter to generate a unstyled table from the source doc tables.

Added specific styling for basic text docs so that we can control things a bit more independently.