replace node-sass with new sass implementation
I had to update node-sass from v4 to v7 to get it to work with ubuntu > 16.04 (which GH actions no longer allows) but then that fails the gatsby build because the gatsby sass plugin expects v4, and to update that we have to udpate our whole gatsby. Plus node-sass is a massive pain and is now deprecated so this needed to happen anyway.